Interface PersoonRepository gemaakt.
