+++ BoeteAPI-backend/src/main/java/nl/fuchsia/controller/PersoonController.java
@@ -15,6 +15,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.util.Optional;
+
 @RestController
 @RequestMapping("/personen")
 public class PersoonController {
@@ -51,7 +53,8 @@ public class PersoonController {
 
     /**
      * verkrijgt  één persoon op bassis van de meegegeven ID nummer.
-     */
+	 * @return
+	 */
     @GetMapping(value = "/{persoonnr}")
     public ResponseEntity<Persoon> getPersoonById(@PathVariable("persoonnr") Integer persoonnr) {
