+++ BoeteAPI-backend/src/main/java/nl/fuchsia/services/PersoonService.java
@@ -1,13 +1,13 @@
 package nl.fuchsia.services;
 
 import java.util.List;
+import java.util.Optional;
 
 import nl.fuchsia.exceptionhandlers.NotFoundException;
 import nl.fuchsia.exceptionhandlers.UniekVeldException;
 import nl.fuchsia.model.Persoon;
 import nl.fuchsia.repository.PersoonRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.TransactionSystemException;
 
@@ -26,7 +26,7 @@ public class PersoonService {
      * @return - Roept de methode getOrmPersonen aan in ormPersoonRepository.
      */
     public List<Persoon> getPersonen() {
-        return persoonRepository.getPersonen();
+        return persoonRepository.findAll();
     }
 
     /**
@@ -36,8 +36,8 @@ public class PersoonService {
      */
     public Persoon addPersoon(Persoon persoon) {
         try {
-            return persoonRepository.addPersoon(persoon);
-        } catch (DataIntegrityViolationException e) {
+            return persoonRepository.save(persoon);
+        } catch (TransactionSystemException e) {
             throw new UniekVeldException("BSN nummer: " + persoon.getBsn() + " bestaat reeds.");
         }
     }
@@ -45,16 +45,13 @@ public class PersoonService {
     /**
      * haalt de persoon per ID (persoonnr) via de OrmPersoonRepository.
      *
-     * @param persoonnr - ID de op te halen persoon.
-     */
+	 * @param persoonnr - ID de op te halen persoon.
+	 * @return
+	 */
     public Persoon getPersoonById(Integer persoonnr) {
-        Persoon persoon = persoonRepository.getPersoonById(persoonnr);
+        Optional<Persoon> persoon = persoonRepository.findById(persoonnr);
 
-        if (persoon == null) {
-            throw new NotFoundException("PersoonNummer: " + persoonnr + " bestaat niet");
-        }
-
-        return persoon;
+        return persoon.orElseThrow(() -> new NotFoundException("PersoonNummer: " + persoonnr + " bestaat niet"));
     }
 
     /**
@@ -67,13 +64,17 @@ public class PersoonService {
 
         try {
             System.out.println(persoonnr);
-            if (persoonRepository.getPersoonById(persoonnr) == null) {
-                throw new NotFoundException("Persoonnummer: " + persoonnr + " bestaat niet!");
-            }
-            persoon.setPersoonnr(persoonnr);
-            persoonRepository.updatePersoonById(persoon);
+			Optional<Persoon> persoon2 = persoonRepository.findById(persoonnr);
 
-        } catch (DataIntegrityViolationException e) {
+			persoon2.orElseThrow(() -> new NotFoundException("PersoonNummer: " + persoonnr + " bestaat niet"));
+
+//			if (persoonRepositoryInt.findById(persoonnr) == null) {
+//                throw new NotFoundException("Persoonnummer: " + persoonnr + " bestaat niet!");
+//            }
+            persoon.setPersoonnr(persoonnr);
+            persoonRepository.save(persoon);
+
+        } catch (TransactionSystemException e) {
             throw new UniekVeldException("BSN nummer: " + persoon.getBsn() + " bestaat reeds.");
         }
